% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevholt.R
\name{bevholt}
\alias{bevholt}
\title{Beverton-Holt}
\usage{
bevholt(S, R, par = list(logRmax = 0, logS50 = 0, logSigma = 0))
}
\arguments{
\item{S}{spawning biomass.}

\item{R}{recruitment.}

\item{par}{initial parameter values.}
}
\value{
Vector of estimated parameter values.
}
\description{
Fit a Beverton-Holt recruitment model.
}
\examples{
bevholt()

}
