% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Fit Model}
\usage{
fitModel(data, parameters = list(logRmax = 0, logS50 = 0, logSigma = 0),
  quiet = TRUE)
}
\arguments{
\item{data}{data frame containing \code{S} and \code{R}, spawning biomass and
recruitment.}

\item{parameters}{vector of initial parameter values.}

\item{quiet}{whether to suppress gradient messages.}
}
\value{
List containing (TMB) \code{model}, (nlminb) \code{fit}, and (SD)
\code{report}.
}
\description{
Fit a Beverton-Holt recruitment model.
}
\note{
The formulation used is \eqn{\hat{R} =
R_\mathrm{max} \frac{S}{S+S_{50}}}{Rhat = Rmax * S / (S + S50)}.

\eqn{R_\mathrm{max}}{Rmax} is maximum recruitment and \eqn{S_{50}}{S50} is
the stock size that produces \eqn{0.50R_\mathrm{max}}{0.50 Rmax}.
}
\examples{
fm <- fitModel(recdata)
fm$fit$par
fm$fit$objective
Rhat <- summary(fm$report)[rownames(summary(fm$report))=="Rhat",]
data.frame(recdata, Rhat=Rhat[,"Estimate"])

}
