% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdata.R
\docType{data}
\name{recdata}
\alias{recdata}
\title{Recruitment Data}
\format{
Data frame containing three columns:
\describe{
\item{\code{Year}}{year.}
\item{\code{S}}{spawning biomass.}
\item{\code{R}}{recruitment.}
}
}
\usage{
recdata
}
\description{
Simulated stock-recruitment data.
}
\examples{
recdata
fitModel(recdata)
}
