% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bevholt.R
\name{bevholt}
\alias{bevholt}
\title{Beverton-Holt Model}
\usage{
bevholt(data, parameters = list(logRmax = 0, logS50 = 0, logSigma = 0),
  quiet = TRUE)
}
\arguments{
\item{data}{data frame containing \code{S} and \code{R}, spawning biomass and
recruitment.}

\item{parameters}{vector of initial parameter values.}

\item{quiet}{whether to suppress gradient messages.}
}
\value{
Vector of estimated parameter values.
}
\description{
Fit a Beverton-Holt recruitment model.
}
\examples{
bevholt(recdata)

}
