% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdata.R
\docType{data}
\name{recdata}
\alias{recdata}
\title{Recruitment Data}
\format{
Data frame containing two columns:
\tabular{ll}{
  \code{S} \tab spawning biomass\cr
  \code{R} \tab recruitment
}
}
\usage{
recdata
}
\description{
Simulated stock-recruitment dataset.
}
\examples{
recdata
bevholt(recdata)
}
